update aria labels, reorganize files, and fix copy
