+++ components/events/calendars/AppleInstructions.vue
@@ -1,7 +1,7 @@
 <template>
   <cv-list ordered>
     <cv-list-item>Open your Apple Calendar</cv-list-item>
-    <cv-list-item>In the menu, select File > New Calendar Subscription</cv-list-item>
+    <cv-list-item>In the menu, select "File" > "New Calendar Subscription"</cv-list-item>
     <cv-list-item>Paste the events subscription link you copied above</cv-list-item>
     <cv-list-item>Click "Subscribe"</cv-list-item>
     <cv-list-item>Click "Ok"</cv-list-item>
